+++ src/com/intellij/idea/plugin/hybris/impex/formatting/ImpexBlock.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -28,14 +28,38 @@ import com.intellij.psi.formatter.common.AbstractBlock
 
 class ImpexBlock(
     node: ASTNode,
-    wrap: Wrap? = null,
-    alignment: Alignment?,
+    wrap: Wrap? = Wrap.createWrap(WrapType.NONE, false),
+    alignment: Alignment? = null,
     private val spacingBuilder: SpacingBuilder,
     private val codeStyleSettings: CodeStyleSettings,
     private val alignmentStrategy: ImpExAlignmentStrategy
 ) : AbstractBlock(node, wrap, alignment) {
 
-    override fun getDebugName() = "ImpEx Block"
+    override fun getDebugName() = when (myNode.elementType) {
+        ImpexTypes.HEADER_LINE -> "Header Line"
+        ImpexTypes.VALUE_LINE -> "Value Line"
+        ImpexTypes.ANY_HEADER_MODE -> "Mode"
+        ImpexTypes.FULL_HEADER_PARAMETER -> "Parameter"
+        ImpexTypes.ANY_HEADER_PARAMETER_NAME -> "Parameter Name"
+        ImpexTypes.FULL_HEADER_TYPE -> "Type"
+        ImpexTypes.HEADER_TYPE_NAME -> "Name"
+        ImpexTypes.SUB_TYPE_NAME -> "Sub-Type"
+        ImpexTypes.PARAMETERS_SEPARATOR -> "Separator"
+        ImpexTypes.FIELD_VALUE_SEPARATOR -> "Separator"
+        ImpexTypes.VALUE_GROUP -> "Value Group"
+        ImpexTypes.VALUE -> "Value"
+        ImpexTypes.LINE_COMMENT -> "Line Comment"
+        ImpexTypes.MODIFIERS -> "Modifiers"
+        ImpexTypes.ANY_ATTRIBUTE_NAME -> "Name"
+        ImpexTypes.ANY_ATTRIBUTE_VALUE -> "Value"
+        ImpexTypes.ASSIGN_VALUE -> "="
+        ImpexTypes.ATTRIBUTE -> "Attribute"
+        ImpexTypes.ATTRIBUTE_SEPARATOR -> ","
+        ImpexTypes.LEFT_SQUARE_BRACKET -> "["
+        ImpexTypes.RIGHT_SQUARE_BRACKET -> "]"
+        else -> "Block"
+    }
+
     override fun isLeaf() = myNode.firstChildNode == null
     override fun getSpacing(child1: Block?, child2: Block): Spacing? = spacingBuilder.getSpacing(this, child1, child2)
     override fun getIndent(): Indent = Indent.getNoneIndent()
@@ -67,13 +91,14 @@ class ImpexBlock(
         return blocks
     }
 
-    private fun isNewBlockToBeMade(currentNode: ASTNode) = isNotWhitespaceOrNewLine(currentNode)
-        && !isCurrentNodeHasParentValue(currentNode)
-
-    private fun isNotWhitespaceOrNewLine(currentNode: ASTNode) = currentNode.elementType != TokenType.WHITE_SPACE
+    private fun isNewBlockToBeMade(currentNode: ASTNode) = currentNode.elementType != TokenType.WHITE_SPACE
         && currentNode.elementType != ImpexTypes.CRLF
-
-    private fun isCurrentNodeHasParentValue(currentNode: ASTNode) = currentNode
-        .treeParent.elementType == ImpexTypes.VALUE
+        && currentNode.elementType != ImpexTypes.HEADER_PARAMETER_NAME
+        && currentNode.elementType != ImpexTypes.ATTRIBUTE_NAME
+        && currentNode.elementType != ImpexTypes.ATTRIBUTE_VALUE
+        && currentNode.elementType != ImpexTypes.HEADER_TYPE
+        && currentNode.elementType != ImpexTypes.VALUE_SUBTYPE
+        && currentNode.treeParent.elementType != ImpexTypes.VALUE
+        && currentNode.treeParent.elementType != ImpexTypes.ANY_ATTRIBUTE_VALUE
 
 }