+++ src/com/intellij/idea/plugin/hybris/impex/formatting/ImpexFormattingModelBuilder.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,9 +18,10 @@
 
 package com.intellij.idea.plugin.hybris.impex.formatting
 
-import com.intellij.formatting.*
-import com.intellij.lang.ASTNode
-import com.intellij.psi.PsiFile
+import com.intellij.formatting.FormattingContext
+import com.intellij.formatting.FormattingModel
+import com.intellij.formatting.FormattingModelBuilder
+import com.intellij.formatting.FormattingModelProvider
 import com.intellij.psi.codeStyle.CodeStyleSettings
 
 class ImpexFormattingModelBuilder : FormattingModelBuilder {
@@ -30,7 +31,6 @@ class ImpexFormattingModelBuilder : FormattingModelBuilder {
         val settings = formattingContext.codeStyleSettings
         val impexBlock = ImpexBlock(
             node = element.node,
-            alignment = Alignment.createAlignment(),
             spacingBuilder = ImpExSpacingBuilder(settings, settings.getCustomSettings(ImpexCodeStyleSettings::class.java)),
             codeStyleSettings = settings,
             alignmentStrategy = getAlignmentStrategy(settings)
@@ -46,5 +46,4 @@ class ImpexFormattingModelBuilder : FormattingModelBuilder {
         else ImpExColumnsAlignmentStrategy()
     }
 
-    override fun getRangeAffectingIndent(file: PsiFile, offset: Int, elementAtOffset: ASTNode?) = null
 }